requirements.txt
pandas==1.3.5
geopy==1.17.0
matplotlib==3.2.2
matplotlib-inline==0.1.3
stellargraph==1.2.1
chardet==3.0.4
charset-normalizer==2.1.0
ipykernel==4.10.1
ipython==5.5.0
ipython-genutils==0.2.0
ipython-sql==0.3.9
ipywidgets==7.7.0
pydot==1.3.0
pyyaml